Added models, Changed enemies
-Added models & textures for enemies.
-Added new behaviours for enemies.
-Changes for Game Manager, fixed issues with level change

+++ Assets/Scripts/UI/UIAmmo.cs
@@ -12,7 +12,7 @@ namespace Tzaik
         [SerializeField] Image ammoImg; 
          
         public void ShowAmmo(Weapon w)
-            => ammoImg.fillAmount = (w != null) ? w.WeaponAmmo.TotalAmmo / w.WeaponAmmo.MaxAmmo : 1;
+            => ammoImg.fillAmount = (w != null) ? (float)w.WeaponAmmo.TotalAmmo / (float)w.WeaponAmmo.MaxAmmo  : 1;
 
     }
 }
